+++ apps/commerce-api/src/main/java/com/loopers/application/inventory/InventoryModifyService.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.inventory;
 
 import org.springframework.stereotype.Service;
