+++ apps/commerce-api/src/main/java/com/loopers/application/member/MemberFacade.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.member;
 
 import java.math.BigDecimal;
