+++ apps/commerce-api/src/main/java/com/loopers/application/member/MemberQueryService.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.member;
 
 import org.springframework.stereotype.Service;
