+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderModifyService.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.order;
 
 import org.springframework.stereotype.Service;
