+++ apps/commerce-api/src/main/java/com/loopers/application/point/PointFacade.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.point;
 
 import java.math.BigDecimal;
