+++ apps/commerce-api/src/main/java/com/loopers/application/product/ProductQueryService.java
@@ -1,4 +1,4 @@
-package com.loopers.application;
+package com.loopers.application.product;
 
 
 import java.util.List;
