+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/member/MemberV1ApiController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.loopers.application.MemberFacade;
+import com.loopers.application.member.MemberFacade;
 import com.loopers.domain.member.Member;
 import com.loopers.domain.member.MemberId;
 import com.loopers.interfaces.api.ApiResponse;
